Exact expected branch lengths
Exact expected branch lengths for Beta-coalescents, and a mixture of a Kingman-coalescent and a Beta-coalescent. This is a CWEB document and can be compiled into (i) a .tex file using cweave for LaTeX compilation into the documentation file, and  (ii) a  .c file using ctangle  for compiling the C code.